projectToDo/requirements.txt
asgiref==3.5.0
black==22.3.0
click==8.1.3
css==0.1
Django==4.0.4
django-crispy-forms==1.14.0
django-jquery==3.1.0
django-rest-multiple-models==2.1.3
djangoajax==3.3
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pystache==0.6.0
pytz==2022.1
sqlparse==0.4.2
static==1.1.1
tomli==2.0.1
whitenoise==6.0.0
